# [chore] discard response body