@atoulme 2024-06-18
Sorry, dup of https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/33587